==> April/Apr05_1_Pandas/p08_replace.py <==
# -*- coding:utf-8 -*-

import pandas as pd

a = pd.read_csv('D:/yu/LNPS.csv', names=['시장이름', '품목', '가격', '시장종류', '구'])
print(a)
print('-----')

# 대형마트 -> 마트 로 바꾸고 싶음
# a = a.replace('대형마트', '마트') # # 필드 구분 안 하고 데이터가 다 바뀜 (필드명은 안 바뀜)

# 특정 필드만 바꾸고 싶으면
a['시장종류'] = a['시장종류'].replace('대형마트', '마트') # 특정 열만
print(a)
print('-----')

# 한꺼번에 바꾸는 첫 번째 스타일
# 마트 -> 일요일에 영업 X, 전통시장 -> 일요일에 영업 O
a['시장종류'] = a['시장종류'].replace(['마트', '전통시장'], ['안 해', '해']) # 한 번에 여러 필드 바꾸기
print(a)
print('-----')

# 한꺼번에 바꾸는 두 번째 스타일
# 안 해 -> 마트, 해 -> 시장
a['시장종류'] = a['시장종류'].replace({'안 해' : '마트', '해' : '시장'}) # dict로도 가능
print(a)
print('-----')

# 한꺼번에 바꾸는 세 번째 스타일
# 마트/시장 -> 물건파는 곳
a['시장종류'] = a['시장종류'].replace(['마트', '시장'], '물건파는 곳')
print(a)
print('-----')

# 필드명 바꾸기
a = a.rename(columns={'시장이름' : '어디', '품목' : '뭐'})
print(a)
